refactor: share mapper metadata parsing and file references
