+++ src/detect.ts
@@ -1,3 +1,9 @@
+import {
+  pyprojectHasToolSection,
+  pythonTomlStringValues,
+  readTomlBracketValue,
+  pythonRequirementName,
+} from "./python-metadata.js";
 import { stripLineComments, stripSwiftComments, stripXmlComments } from "./source-comments.js";
 import { lstat, readFile, readdir } from "node:fs/promises";
 import { join, posix } from "node:path";
@@ -645,15 +651,6 @@ async function pythonRunner(root: string): Promise<string | null> {
   return null;
 }
 
-async function pyprojectHasToolSection(root: string, tool: string): Promise<boolean> {
-  if (!(await pathExists(join(root, "pyproject.toml")))) {
-    return false;
-  }
-  const source = await readFile(join(root, "pyproject.toml"), "utf8");
-  const escaped = tool.replace(/[.*+?^${}()|[\]\\]/gu, "\\$&");
-  return new RegExp(`^\\s*\\[\\[?tool\\.${escaped}(?:\\.|\\])`, "mu").test(source);
-}
-
 function pythonRunCommand(runner: string | null, command: string): string {
   if (runner === "uv") {
     return `uv run ${command}`;
@@ -918,72 +915,6 @@ function pythonTomlArrayValues(source: string): string[] {
   return pythonTomlStringValues(source);
 }
 
-function pythonTomlStringValues(source: string): string[] {
-  const values: string[] = [];
-  let quote: string | null = null;
-  let value = "";
-  let escaped = false;
-  for (let index = 0; index < source.length; index += 1) {
-    const char = source[index];
-    if (quote !== null) {
-      if (escaped) {
-        value += char;
-        escaped = false;
-      } else if (char === "\\" && quote === '"') {
-        escaped = true;
-      } else if (char === quote) {
-        values.push(value);
-        quote = null;
-        value = "";
-      } else {
-        value += char;
-      }
-      continue;
-    }
-    if (char === "#") {
-      const nextNewline = source.indexOf("\n", index + 1);
-      if (nextNewline === -1) {
-        break;
-      }
-      index = nextNewline;
-    } else if (char === '"' || char === "'") {
-      quote = char;
-      value = "";
-    }
-  }
-  return values;
-}
-
-function readTomlBracketValue(source: string, bracketIndex: number): string {
-  let depth = 0;
-  let quote: string | null = null;
-  let escaped = false;
-  for (let index = bracketIndex; index < source.length; index += 1) {
-    const char = source[index];
-    if (quote !== null) {
-      if (escaped) {
-        escaped = false;
-      } else if (char === "\\") {
-        escaped = true;
-      } else if (char === quote) {
-        quote = null;
-      }
-      continue;
-    }
-    if (char === '"' || char === "'") {
-      quote = char;
-    } else if (char === "[") {
-      depth += 1;
-    } else if (char === "]") {
-      depth -= 1;
-      if (depth === 0) {
-        return source.slice(bracketIndex, index + 1);
-      }
-    }
-  }
-  return source.slice(bracketIndex);
-}
-
 function pythonRequirementNames(source: string): string[] {
   return source
     .split("\n")
@@ -1035,15 +966,6 @@ function addPythonRequirementNames(names: Set<string>, value: string): void {
       names.add(name);
     }
   }
-}
-
-function pythonRequirementName(value: string): string | null {
-  const trimmed = value.trim().replace(/^["']|["']$/gu, "");
-  if (trimmed.length === 0 || trimmed.startsWith("#") || trimmed.startsWith("-")) {
-    return null;
-  }
-  const match = /^([A-Za-z0-9_.-]+)/u.exec(trimmed);
-  return match?.[1]?.toLowerCase().replace(/_/gu, "-") ?? null;
 }
 
 async function containsPythonTestFile(root: string, maxDepth: number): Promise<boolean> {
