+++ src/mappers/node.ts
@@ -5,6 +5,7 @@ import { pathExists } from "../fs.js";
 import { rubyDependencyNames, rubyGemspecPaths, stripRubyComments } from "../ruby.js";
 import { partitionFileGroups } from "./grouping.js";
 import {
+  uniqueFileRefs,
   normalize,
   packageKind,
   packageTrustBoundaries,
@@ -671,17 +672,4 @@ function isExtensionPackage(info: PackageInfo): boolean {
 
 function isNodeTestPath(path: string): boolean {
   return /\.(test|spec)\.(ts|tsx|js|jsx|mts|cts|mjs|cjs)$/u.test(path);
-}
-
-function uniqueFileRefs(refs: SeedFileRef[]): SeedFileRef[] {
-  const seen = new Set<string>();
-  const output: SeedFileRef[] = [];
-  for (const ref of refs) {
-    if (seen.has(ref.path)) {
-      continue;
-    }
-    seen.add(ref.path);
-    output.push(ref);
-  }
-  return output;
 }
